pyproject.toml
[project]
name = "vkhack"
version = "0.1.0"
description = ""
authors = [
    {name = "dreadew",email = "p74ur@yandex.ru"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pika (>=1.3.2,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "fastapi (>=0.117.1,<0.118.0)",
    "prometheus-client (>=0.23.1,<0.24.0)",
    "uvicorn[standard] (>=0.36.0,<0.37.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "celery[asyncio] (>=5.5.3,<6.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "sqlalchemy-trino (>=0.5.0,<0.6.0)",
    "black (>=25.9.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | __pycache__
  | migrations
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src", "tests"]
known_third_party = ["fastapi", "pydantic", "sqlalchemy", "celery", "pytest"]